+++ lib/zendesk2/category.rb
