+++ lib/zendesk2/client.rb
@@ -6,22 +6,22 @@ require 'zendesk2/collection'
 require 'zendesk2/help_center'
 
 require 'zendesk2/requests/search'
-require 'zendesk2/models/audit_event'
-require 'zendesk2/models/category'
+require 'zendesk2/audit_event'
+require 'zendesk2/category'
 require 'zendesk2/collections/categories'
 require 'zendesk2/requests/create_category'
 require 'zendesk2/requests/destroy_category'
 require 'zendesk2/requests/get_categories'
 require 'zendesk2/requests/get_category'
 require 'zendesk2/requests/update_category'
-require 'zendesk2/models/forum'
+require 'zendesk2/forum'
 require 'zendesk2/collections/forums'
 require 'zendesk2/requests/create_forum'
 require 'zendesk2/requests/destroy_forum'
 require 'zendesk2/requests/get_forums'
 require 'zendesk2/requests/get_forum'
 require 'zendesk2/requests/update_forum'
-require 'zendesk2/models/group'
+require 'zendesk2/group'
 require 'zendesk2/collections/groups'
 require 'zendesk2/requests/create_group'
 require 'zendesk2/requests/destroy_group'
@@ -29,7 +29,7 @@ require 'zendesk2/requests/get_groups'
 require 'zendesk2/requests/get_group'
 require 'zendesk2/requests/update_group'
 require 'zendesk2/requests/get_assignable_groups'
-require 'zendesk2/models/user'
+require 'zendesk2/user'
 require 'zendesk2/collections/users'
 require 'zendesk2/requests/create_user'
 require 'zendesk2/requests/destroy_user'
@@ -41,7 +41,7 @@ require 'zendesk2/requests/get_current_user'
 require 'zendesk2/requests/mark_user_identity_primary'
 require 'zendesk2/requests/get_user_memberships'
 require 'zendesk2/requests/get_user_organizations'
-require 'zendesk2/models/ticket'
+require 'zendesk2/ticket'
 require 'zendesk2/collections/tickets'
 require 'zendesk2/requests/create_ticket'
 require 'zendesk2/requests/destroy_ticket'
@@ -50,34 +50,34 @@ require 'zendesk2/requests/get_ticket'
 require 'zendesk2/requests/update_ticket'
 require 'zendesk2/requests/get_requested_tickets'
 require 'zendesk2/requests/get_ccd_tickets'
-require 'zendesk2/models/ticket_audit'
+require 'zendesk2/ticket_audit'
 require 'zendesk2/collections/ticket_audits'
 require 'zendesk2/requests/get_ticket_audits'
 require 'zendesk2/requests/get_ticket_audit'
-require 'zendesk2/models/ticket_field'
+require 'zendesk2/ticket_field'
 require 'zendesk2/collections/ticket_fields'
 require 'zendesk2/requests/create_ticket_field'
 require 'zendesk2/requests/destroy_ticket_field'
 require 'zendesk2/requests/get_ticket_fields'
 require 'zendesk2/requests/get_ticket_field'
 require 'zendesk2/requests/update_ticket_field'
-require 'zendesk2/models/topic'
+require 'zendesk2/topic'
 require 'zendesk2/collections/topics'
 require 'zendesk2/requests/create_topic'
 require 'zendesk2/requests/destroy_topic'
 require 'zendesk2/requests/get_topics'
 require 'zendesk2/requests/get_topic'
 require 'zendesk2/requests/update_topic'
-require 'zendesk2/models/topic_comment'
+require 'zendesk2/topic_comment'
 require 'zendesk2/collections/topic_comments'
 require 'zendesk2/requests/create_topic_comment'
 require 'zendesk2/requests/destroy_topic_comment'
 require 'zendesk2/requests/get_topic_comments'
 require 'zendesk2/requests/get_topic_comment'
-require 'zendesk2/models/ticket_comment'
+require 'zendesk2/ticket_comment'
 require 'zendesk2/collections/ticket_comments'
 require 'zendesk2/requests/get_ticket_comments'
-require 'zendesk2/models/organization'
+require 'zendesk2/organization'
 require 'zendesk2/collections/organizations'
 require 'zendesk2/requests/create_organization'
 require 'zendesk2/requests/destroy_organization'
@@ -89,28 +89,28 @@ require 'zendesk2/requests/get_organization_tickets'
 require 'zendesk2/requests/get_organization_by_external_id'
 require 'zendesk2/requests/get_organization_memberships'
 require 'zendesk2/requests/search_organization'
-require 'zendesk2/models/user_field'
+require 'zendesk2/user_field'
 require 'zendesk2/collections/user_fields'
 require 'zendesk2/requests/create_user_field'
 require 'zendesk2/requests/destroy_user_field'
 require 'zendesk2/requests/get_user_fields'
 require 'zendesk2/requests/get_user_field'
 require 'zendesk2/requests/update_user_field'
-require 'zendesk2/models/user_identity'
+require 'zendesk2/user_identity'
 require 'zendesk2/collections/user_identities'
 require 'zendesk2/requests/create_user_identity'
 require 'zendesk2/requests/destroy_user_identity'
 require 'zendesk2/requests/get_user_identities'
 require 'zendesk2/requests/get_user_identity'
 require 'zendesk2/requests/update_user_identity'
-require 'zendesk2/models/membership'
+require 'zendesk2/membership'
 require 'zendesk2/collections/memberships'
 require 'zendesk2/requests/create_membership'
 require 'zendesk2/requests/destroy_membership'
 require 'zendesk2/requests/get_memberships'
 require 'zendesk2/requests/get_membership'
 require 'zendesk2/requests/mark_membership_default'
-require 'zendesk2/models/view'
+require 'zendesk2/view'
 require 'zendesk2/collections/views'
 require 'zendesk2/requests/create_view'
 require 'zendesk2/requests/destroy_view'
