+++ lib/zendesk2/group.rb
