+++ lib/zendesk2/ticket_comment.rb
