+++ lib/zendesk2/ticket_comment_privacy_change.rb
