+++ lib/zendesk2/ticket_metric.rb
