+++ lib/zendesk2/topic_comment.rb
