+++ lib/zendesk2/user_identity.rb
