+++ lib/zendesk2/view.rb
